refactor(Command): Command 객체 view 패키지로 이동
- Command는 사용자 입력에 대해 작동 흐름을 보여주는 view이다.
