+++ src/main/java/view/Command.java
@@ -1,4 +1,4 @@
-package model;
+package view;
 
 import java.util.Arrays;
